Security subsystem, user authentication, Ext2 write/create support, bug fixes.
- Added DAC security checks and permissions checking across VFS and syscalls
- Implemented user accounts, PBKDF2-HMAC-SHA256 password hashing, and /etc/shadow persistence
- Added interactive login prompt (hidden password input) and useradd/login/su shell commands
- Added Ext2 inode allocation, directory entry creation, and disk node creation
- Fixed hardcoded sector calculation in Ext2 block reads/writes to use sectors_per_block
- Fixed fork allocating only 4KB instead of full 16KB contiguous kernel stack
- Fixed wait syscall memory leak by freeing contiguous frames on child stack cleanup
- Fixed exception and keyboard ISRs corrupting segment registers across privilege boundaries
- Added restore_context_and_iret assembly helper and CR2 reporting for page faults
- Added contiguous physical page allocation and kernel PDE sync across address spaces
- Added UART COM1 serial mirroring for text output and enabled serial stdio in QEMU

+++ src/syscalls/time.rs
@@ -5,10 +5,10 @@ use crate::startup_config::power::CONFIG_HZ;
 pub(super) unsafe fn syscall_nanosleep(regs: *mut ContextFrame) {
     unsafe {
         let sleep_ms = (*regs).arg1();
-        let ticks_to_wait = (sleep_ms as u64) * (CONFIG_HZ as u64) / 1000;
-        let wakeup_tick = timer::get_ticks() + ticks_to_wait;
+        let ticks_to_wait = (sleep_ms) * (CONFIG_HZ) / 1000;
+        let wakeup_tick = timer::get_ticks() as u64 + ticks_to_wait as u64;
         let task = sched::current().as_mut().unwrap();
-        task.wakeup_time = wakeup_tick;
+        task.wakeup_time = wakeup_tick as u64;
         task.state = ProcessState::Sleeping;
 
         while sched::current().as_ref().unwrap().state == ProcessState::Sleeping {
